Rename legacy files and add initial Python implementation for NetPhorest kinase substrate prediction
